+++ nn_dataflow/core/resource.py
@@ -33,6 +33,7 @@ RESOURCE_LIST = ['proc_region',
                  'size_regf',
                  'array_bus_width',
                  'dram_bandwidth',
+                 'src_data_region_final_',
                 ]
 
 class Resource(namedtuple('Resource', RESOURCE_LIST)):
@@ -43,6 +44,10 @@ class Resource(namedtuple('Resource', RESOURCE_LIST)):
     '''
 
     def __new__(cls, *args, **kwargs):
+        if len(args) <= RESOURCE_LIST.index('src_data_region_final_'):
+            kwargs = kwargs.copy()
+            kwargs.setdefault('src_data_region_final_', None)
+
         ntp = super(Resource, cls).__new__(cls, *args, **kwargs)
 
         if not isinstance(ntp.proc_region, NodeRegion):
@@ -62,6 +67,9 @@ class Resource(namedtuple('Resource', RESOURCE_LIST)):
                             'a NodeRegion instance.')
         if not isinstance(ntp.dst_data_region, NodeRegion):
             raise TypeError('Resource: dst_data_region must be '
+                            'a NodeRegion instance.')
+        if not isinstance(ntp.src_data_region_final, NodeRegion):
+            raise TypeError('Resource: src_data_region_final must be '
                             'a NodeRegion instance.')
 
         if not isinstance(ntp.dim_array, PhyDim2):
@@ -85,4 +93,10 @@ class Resource(namedtuple('Resource', RESOURCE_LIST)):
             raise ValueError('Resource: dram_bandwidth must be positive.')
 
         return ntp
+
+    @property
+    def src_data_region_final(self):
+        ''' Final source data region after resolving forwarding and sharing. '''
+        return self.src_data_region_final_ if self.src_data_region_final_ \
+                else self.src_data_region